automl/frameworks/hyperoptsklearn/py_requirements.txt
hyperopt-sklearn
networkx==1.11